refactor: extract visit_c_expr to new module
